[progress-chart] Split SetList composable into its own file
