+++ app/src/main/java/com/litus_animae/refitted/compose/exercise/Main.kt
@@ -106,100 +106,3 @@ fun Exercise(
     }
   }
 }
-
-@Composable
-private fun SetRecordList(flow: Flow<PagingData<SetRecord>>) {
-  val records = flow.collectAsLazyPagingItems()
-  LazyColumn {
-    item {
-      Row(
-        Modifier
-          .fillMaxWidth()
-          .background(MaterialTheme.colors.primary)
-          .padding(start = 10.dp, top = 10.dp, bottom = 10.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-      ) {
-        // TODO localize
-        Text(
-          "Set History", style = MaterialTheme.typography.h6, color = contentColorFor(
-            backgroundColor = MaterialTheme.colors.primary
-          )
-        )
-        Icon(
-          Icons.Default.Refresh,
-          // TODO localize
-          "refresh",
-          modifier = Modifier
-            .clickable {
-              records.refresh()
-            }
-            .padding(start = 10.dp, end = 10.dp),
-          tint = contentColorFor(backgroundColor = MaterialTheme.colors.primary))
-      }
-    }
-
-    if (records.loadState.refresh is LoadState.Loading) {
-      item {
-        Row(Modifier.fillMaxWidth()) {
-          LoadingView()
-        }
-      }
-    } else {
-      val dateFormat = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.MEDIUM)
-        .withZone(ZoneId.systemDefault())
-      item {
-        Row(
-          Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 10.dp, vertical = 15.dp),
-          horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-          Text(
-            stringResource(id = R.string.date),
-            style = MaterialTheme.typography.button
-          )
-          Text(
-            stringResource(id = R.string.reps_label),
-            style = MaterialTheme.typography.button
-          )
-          val weightLabel = stringResource(id = R.string.weight_label)
-          val weightUnits = stringResource(id = R.string.lbs)
-          Text(
-            "$weightLabel ($weightUnits)",
-            style = MaterialTheme.typography.button
-          )
-        }
-      }
-
-      items(
-        count = records.itemCount,
-        key = records.itemKey { it.completed }
-      ) { index ->
-        val record = records[index]
-        if (record != null) {
-          Row(
-            Modifier
-              .fillMaxWidth()
-              .padding(horizontal = 10.dp, vertical = 15.dp),
-            horizontalArrangement = Arrangement.SpaceBetween
-          ) {
-            Text(
-              dateFormat.format(record.completed),
-              style = MaterialTheme.typography.button
-            )
-            Text(
-              record.reps.toString(),
-              style = MaterialTheme.typography.button
-            )
-            Text(
-              String.format("%.1f", record.weight),
-              style = MaterialTheme.typography.button
-            )
-          }
-          Divider()
-        }
-      }
-    }
-  }
-}
